# swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.2",
        "title": "Kidslike v2 API",
        "description": "This is a server for a Kidslike app. Github repository: https://github.com/vshynder/kidslike-back",
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        }
    },
    "host": "kidslike-back-end.herokuapp.com/",
    "schemes": [
        "https"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
...
    "paths": {
...
        "api/presents/": {
            "get": {
                "tags": [
                    "Подарки"
                ],
                "summary": "Получение подарков",
                "description": "",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "type": "string",
                        "description": "accessToken",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": "application/json",
                        "schema": {
                            "$ref": "#/components/schemas/getPresents"
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Подарки"
                ],
                "summary": "Добавление подарка",
                "description": "",
                "parameters": [
                    {
                        "name": "Request body",
                        "in": "body",
                        "required": true,
                        "description": "Объект данных для добавления подарка",
                        "schema": {
                            "$ref": "#/components/schemas/addPresent"
                        }
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "type": "string",
                        "description": "accessToken",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/components/schemas/present"
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
        "api/presents/:presentId": {
            "patch": {
                "tags": [
                    "Подарки"
                ],
                "summary": "Обновление подарка",
                "description": "Update child's present",
                "parameters": [
                    {
                        "name": "Present's Id",
                        "in": "path",
                        "type": "string",
                        "required": true,
                        "description": "Present's Id"
                    },
                    {
                        "name": "Present",
                        "in": "body",
                        "type": "object",
                        "required": true,
                        "description": "Data object to update present",
                        "schema": {
                            "$ref": "#/components/schemas/updatePresent"
                        }
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "type": "string",
                        "description": "accessToken",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/components/schemas/present"
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Подарки"
                ],
                "summary": "Удаление подарка",
                "description": "",
                "parameters": [
                    {
                        "name": "Present's Id",
                        "in": "path",
                        "type": "string",
                        "required": true,
                        "description": "Present's Id"
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "type": "string",
                        "description": "accessToken",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Present deleted successful"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "present": {
                "type": "object",
                "properties": {
                    "_id": {
                        "type": "string"
                    },
                    "title": {
                        "type": "string"
                    },
                    "childId": {
                        "type": "string"
                    },
                    "reward": {
                        "type": "number"
                    },
                    "image": {
                        "type": "string"
                    },
                    "dateCreated": {
                        "type": "string"
                    }
                }
            },
            "addPresent": {
                "type": "object",
                "required:": [
                    "title",
                    "childId",
                    "reward"
                ],
                "properties": {
                    "title": {
                        "type": "string"
                    },
                    "childId": {
                        "type": "string"
                    },
                    "reward": {
                        "type": "number"
                    }
                }
            },
            "getPresents": {
                "type": "array",
                "description": "Array of presents",
                "items": {
                    "type": "object",
                    "properties": {
                        "_id": {
                            "type": "string"
                        },
                        "title": {
                            "type": "string"
                        },
                        "childId": {
                            "type": "string"
                        },
                        "reward": {
                            "type": "number"
                        },
                        "image": {
                            "type": "string"
                        },
                        "dateCreated": {
                            "type": "string"
                        }
                    }
                }
            },
...
            "updatePresent": {
                "type": "object",
                "properties": {
                    "title": {
                        "type": "string"
                    },
                    "reward": {
                        "type": "number"
                    },
                    "image": {}
                }
            }
        }
    }
}